+++ src/main/java/com/intellecteu/onesource/integration/repository/entity/onesource/TradeAgreementEntity.java
@@ -2,8 +2,8 @@ package com.intellecteu.onesource.integration.repository.entity.onesource;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.intellecteu.onesource.integration.model.onesource.CurrencyCd;
 import com.intellecteu.onesource.integration.model.enums.ProcessingStatus;
+import com.intellecteu.onesource.integration.model.onesource.CurrencyCd;
 import com.intellecteu.onesource.integration.model.onesource.SettlementType;
 import com.intellecteu.onesource.integration.model.onesource.TermType;
 import jakarta.persistence.CascadeType;
@@ -57,7 +57,7 @@ public class TradeAgreementEntity {
     @Enumerated(value = EnumType.STRING)
     private CurrencyCd billingCurrency;
     @Column(name = "dividend_rate")
-    private Integer dividendRatePct;
+    private Double dividendRatePct;
     @Column(name = "trade_date", columnDefinition = "DATE")
     private LocalDate tradeDate;
     @Column(name = "term_type")
